Refactor app file into modules
