+++ lib/common.py
@@ -16,9 +16,11 @@ dict_Change_Effect_State = 'ChangeEffectState'
 dict_change_parameter = 'change_parameter'
 dict_change_pedal_preset = 'change_pedal_preset'
 dict_comp = 'COMP'
+dict_connect = 'connect'
 dict_connection_lost = 'connection_lost'
 dict_connection_message = 'connection-message'
 dict_connection_success = 'connection-success'
+dict_debug_logging = 'debug_logging'
 dict_db_id = 'db_id'
 dict_delay = 'DELAY'
 dict_drive = 'DRIVE'
@@ -27,6 +29,8 @@ dict_Effect = 'Effect'
 dict_effect_name = 'effect_name'
 dict_effect_type = 'effect_type'
 dict_enabled = 'enabled'
+dict_error = 'error'
+dict_false = 'false'
 dict_gate = 'GATE'
 dict_GuitarEQ = 'GuitarEQ6'
 dict_id = 'id'
@@ -63,6 +67,7 @@ dict_New_Preset = 'NewPreset'
 dict_new_effect = 'new_effect'
 dict_New_Effect = 'NewEffect'
 dict_Off = 'Off'
+dict_ok = 'ok'
 dict_old_effect = 'old_effect'
 dict_Old_Effect = 'OldEffect'
 dict_On = 'On'
@@ -86,6 +91,7 @@ dict_show_hide_pedal = 'show_hide_pedal'
 dict_state = 'state'
 dict_switch_parameter = 'switch_parameter'
 dict_turn_on_off = 'turn_on_off'
+dict_url = 'url'
 dict_update_effect = 'update-effect'
 dict_update_onoff = 'update-onoff'
 dict_update_parameter = 'update-parameter'
